\name{leptoSRMedgespp}
\alias{leptoSRMedgespp}
\docType{data}
\title{SRM quantification: edges informationu.}
\description{
  This dataframe holds the edge information linking \code{\link{leptoSRMpeptides}} to \code{\link{leptoSRMproteins}}}
\usage{data(leptoSRM)}
\format{
  data frame with 151 edge definitions (=151 rows)
}
\source{Ludwig et al., see references}
\references{
Ludwig et al. Estimation of absolute protein quantities of unlabeled samples by selected reaction monitoring mass spectrometry (2011) \emph{Molecular & Cellular Proteomics}
}
\keyword{datasets}
