\name{leptoSRMpeptides}
\alias{leptoSRMpeptides}
\docType{data}
\title{SRM quantification: peptide example data}
\description{
  This dataframe holds the peptide SRM measurements.}
\usage{data(leptoSRM)}
\format{
  data frame with 151 distinct peptides (=151 rows)
}
\source{Ludwig et al., see references}
\references{
Ludwig et al. Estimation of absolute protein quantities of unlabeled samples by selected reaction monitoring mass spectrometry (2011) \emph{Molecular & Cellular Proteomics}
}
\keyword{datasets}
