\name{leptoSRMproteins}
\alias{leptoSRMproteins}
\docType{data}
\title{SRM quantification: protein example data}
\description{
  This dataframe holds the proteins matching the the peptides in \code{\link{leptoSRMpeptides}}.}
\usage{data(leptoSRM)}
\format{
  data frame with 39 distinct proteins (=39 rows)
}
\source{Ludwig et al., see references}
\references{
Ludwig et al. Estimation of absolute protein quantities of unlabeled samples by selected reaction monitoring mass spectrometry (2011) \emph{Molecular & Cellular Proteomics}
}
\keyword{datasets}
